Fix mechanical jb warnings and CS8602 errors in TimelineSet fast-mixed paths
Structural hoisted-seed fix for FastMixedForward/Backward (no added
branches); span-exact removals for RedundantCast/UnsafeContext (with
pointer-signature restorations), NameQualifier, UsingDirective,
ExplicitArrayCreation, SuppressNullableWarningExpression,
ArgumentDefaultValue, EmptyStatement, RedundantJumpStatement;
CheckNamespace for Numbers/PairHandles; InconsistentNaming renames in
tests/tools/benchmarks.

+++ benchmarks/Numbers/Domain.cs
@@ -177,7 +177,7 @@ internal static class Corpus
         var edges = new uint[ClipsPerTrack];
         for (var j = 0; j < ClipsPerTrack; j++)
         {
-            var ideal = (uint)((j + 1) * (long)Duration / ClipsPerTrack);
+            var ideal = (uint)((j + 1) * Duration / ClipsPerTrack);
             var edge = j == ClipsPerTrack - 1 ? Duration : ideal - 200 + rng.NextUInt32() % 400;
             if (j > 0 && edge <= edges[j - 1]) edge = edges[j - 1] + 1;
             edges[j] = edge;
